+++ crates/apollo-smith/src/lib.rs
@@ -20,7 +20,8 @@ pub(crate) mod ty;
 pub(crate) mod union;
 pub(crate) mod variable;
 
-use std::{collections::HashMap, fmt::Debug};
+use indexmap::IndexMap;
+use std::fmt::Debug;
 
 use arbitrary::Unstructured;
 
@@ -88,9 +89,9 @@ pub struct DocumentBuilder<'a> {
     // A stack to set current ObjectTypeDef
     pub(crate) stack: Vec<Box<dyn StackedEntity>>,
     // Useful to keep the same arguments for a specific field
-    pub(crate) chosen_arguments: HashMap<Name, Vec<Argument>>,
+    pub(crate) chosen_arguments: IndexMap<Name, Vec<Argument>>,
     // Useful to keep the same aliases for a specific field name
-    pub(crate) chosen_aliases: HashMap<Name, Name>,
+    pub(crate) chosen_aliases: IndexMap<Name, Name>,
 }
 
 impl<'a> Debug for DocumentBuilder<'a> {
@@ -126,8 +127,8 @@ impl<'a> DocumentBuilder<'a> {
             union_type_defs: Vec::new(),
             input_object_type_defs: Vec::new(),
             stack: Vec::new(),
-            chosen_arguments: HashMap::new(),
-            chosen_aliases: HashMap::new(),
+            chosen_arguments: IndexMap::new(),
+            chosen_aliases: IndexMap::new(),
         };
 
         for _ in 0..builder.u.int_in_range(1..=50)? {
@@ -200,8 +201,8 @@ impl<'a> DocumentBuilder<'a> {
             union_type_defs: document.union_type_definitions,
             input_object_type_defs: document.input_object_type_definitions,
             stack: Vec::new(),
-            chosen_arguments: HashMap::new(),
-            chosen_aliases: HashMap::new(),
+            chosen_arguments: IndexMap::new(),
+            chosen_aliases: IndexMap::new(),
         };
 
         Ok(builder)